initial paging init implementation (might need hhdm to offset kmalloc calls in map_page())

+++ kernel/include/kernel/system.h
@@ -35,8 +35,8 @@ typedef struct lm_sysseg_desc {
   uint32_t reserved;
 } lm_sysseg_desc_t;
 
-extern uint32_t *page_directory;
-extern uint32_t *first_page_table;
+extern uint64_t *page_directory;
+extern uint64_t *first_page_table;
 
 void gdt_init();
 void paging_init();
